+++ src/main/java/com/dku/council/domain/admin/controller/AdminPageController.java
@@ -1,4 +1,4 @@
-package com.dku.council.admin.page.controller;
+package com.dku.council.domain.admin.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
