+++ src/main/java/com/dku/council/domain/admin/controller/TagPageController.java
@@ -1,7 +1,7 @@
-package com.dku.council.admin.page.controller;
+package com.dku.council.domain.admin.controller;
 
-import com.dku.council.admin.page.dto.TagPageDto;
-import com.dku.council.admin.page.service.TagPageService;
+import com.dku.council.domain.admin.dto.TagPageDto;
+import com.dku.council.domain.admin.service.TagPageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
