+++ src/main/resources/templates/fragment/layout.html
@@ -1,0 +1,4 @@
+<!DOCTYPE html>
+<html>
+
+</html>
