+++ src/main/resources/templates/page/admin/index.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <link rel="stylesheet"
-        th:href="@{/css/bootstrap.min.css}">
+        href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css">
 </head>
 <body>
 <div class="container" style="max-width: 1500px">
