+++ src/main/resources/templates/page/rental/items.html
