+++ src/main/resources/templates/page/rental/rental.html
